Tidy up errors package
